Revised many files in recon framework to remove unnesecessary nexus dependencies. ModuleConfigTest project compiles now. Basic modulenamemanager stucture implemented, no analytics.
